Implemented fade in to vision + text during experiment pause
Uses the status message UI elements in order to display text on the screen when the participant needs to press a button to advance the experiment. There is also a fade in system that should fade in the participants view from the initial blackout portion. This fade in duration is adjustable in the HUD script located in the screen space object, and does not affect the experiment timing as it runs via a coroutine.